feat: initial setup
