--- packages/components/src/_components/dialog/dialog.stories.ts
@@ -1,26 +1,0 @@
-import '../../solid-components';
-import { storybookDefaults, storybookTemplate } from '../../../scripts/storybook/helper';
-import { withActions } from '@storybook/addon-actions/decorator';
-
-const { argTypes, args, parameters } = storybookDefaults('sd-dialog');
-const { generateTemplate } = storybookTemplate('sd-dialog');
-
-export default {
-  title: 'Components/sd-dialog',
-  component: 'sd-dialog',
-  args,
-  argTypes,
-  parameters: {...parameters},
-  decorators: [withActions] as any
-};
-
-
-/**
- * Default: This shows sd-dialog in its default state.
- */
-
-export const Default = {
-  render: (args: any) => {
-    return generateTemplate({ args });
-  }
-};
