+++ packages/components/src/components/dialog/dialog.stories.ts
@@ -1,0 +1,30 @@
+import '../../solid-components';
+import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../../scripts/storybook/helper';
+// import { withActions } from '@storybook/addon-actions/decorator';
+
+const { overrideArgs } = storybookHelpers('sd-dialog');
+const { argTypes, parameters } = storybookDefaults('sd-dialog');
+const { generateTemplate } = storybookTemplate('sd-dialog');
+
+export default {
+  title: 'Components/sd-dialog',
+  component: 'sd-dialog',
+  args: overrideArgs({
+    type: 'slot',
+    name: 'default',
+    value: `<div class="slot slot--border slot--text h-full">Main slot</div>`
+  }),
+  argTypes,
+  parameters: { ...parameters }
+  // decorators: [withActions] as any
+};
+
+/**
+ * Default: This shows sd-dialog in its default state.
+ */
+
+export const Default = {
+  render: (args: any) => {
+    return generateTemplate({ args });
+  }
+};
